feat: Verilator Test Suite (#7)
* Using boost test and verilator to run RISCV module level and system level tests.

* Removing icarus/slang for now

* Moving TB code out of the rtl sources

* Adding additional module level tests

* Adding additional system level tests

* Adding print for running on HOST machine to verify test works properly

* All tests passing now. Found bug in shift instruction from imm_gen.sv and Byte/Half load store instructions are not implemented so memcpy was failing when not word alligned.

+++ test/gcd/compile.sh
@@ -1,26 +1,19 @@
 #!/bin/bash
 
-DOCKER_IMAGE="ubuntu:riscv_compile"
-DOCKER_LAUNCHER="docker run --rm -it --user $(id -u):$(id -g) -v $PWD:/workspace:z $DOCKER_IMAGE"
+mkdir -p build
 
-mkdir build
+riscv64-unknown-elf-gcc -march=rv32i -mabi=ilp32 -nostdlib -ffreestanding -Tld/memory_map.ld -o build/gcd.elf ld/entry.s src/gcd.c
+if [ $? -ne 0 ]; then
+  exit -1
+fi
 
-$DOCKER_LAUNCHER \
-  riscv64-unknown-elf-gcc -march=rv32i -mabi=ilp32 -nostdlib -ffreestanding -Tld/memory_map.ld -o build/gcd.elf ld/entry.s src/gcd.c
-
-$DOCKER_LAUNCHER \
-  riscv64-unknown-elf-gcc -march=rv32i -mabi=ilp32 -nostdlib -ffreestanding -Tld/memory_map.ld -o build/gcd.s src/gcd.c -S
+riscv64-unknown-elf-gcc -march=rv32i -mabi=ilp32 -nostdlib -ffreestanding -Tld/memory_map.ld -o build/gcd.s src/gcd.c -S
 
 if [ -e build/gcd.dump ]; then
   rm -f build/gcd.dump
 fi
 
-$DOCKER_LAUNCHER \
-  riscv64-unknown-elf-objdump -d build/gcd.elf > build/gcd.dump
+riscv64-unknown-elf-objdump -d build/gcd.elf > build/gcd.dump
 
 # Format for init file in icarus Verilog
 hexdump -v -e '/1 "%02X "' build/gcd.elf > gcd.ini
-
-../pad.py gcd.ini 16
-
-gcc -DLOCAL_COMPILE src/gcd.c -o build/gcd_debug
